+++ tests/__fixtures__/data/vertical/results.ts
